add authorization for campground EDIT form and DELETE routes

+++ routes/campgrounds.js
@@ -43,15 +43,22 @@ router.get('/:id', catchAsync(async (req, res) => {
 }))
 
 router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
-  const campground = await Campground.findById(req.params.id);
+  const {id} = req.params;
+  const campground = await Campground.findById(id);
+  if (!campground) {
+    req.flash('error', 'Cannot find that campground!');
+    return res.redirect('/campgrounds');
+  }
+  if (!campground.author.equals(req.user._id)) {
+    req.flash('error', 'You do not have permission to do that');
+    return res.redirect(`/campgrounds/${id}`)
+  }
   res.render('campgrounds/edit', {campground});
 }))
 
 router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
   const {id} = req.params;
   const campground = await Campground.findById(id);
-  console.log('id: ', id);
-  console.log('campground: ', campground);
   if (!campground.author.equals(req.user._id)) {
     req.flash('error', 'You do not have permission to do that');
     return res.redirect(`/campgrounds/${id}`)
@@ -63,7 +70,12 @@ router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) =
 }))
 
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
-  const { id } = req.params;
+  const {id} = req.params;
+  const campground = await Campground.findById(id);
+  if (!campground.author.equals(req.user._id)) {
+    req.flash('error', 'You do not have permission to do that');
+    return res.redirect(`/campgrounds/${id}`)
+  }
   await Campground.findByIdAndDelete(id);
   req.flash('success', 'Successfully deleted campground');
   res.redirect('/campgrounds');
